Fix MSL classification when there are NAs in data
When there was NA in the data file, these were taken as NA in the dataMIX. After, when assessing
fragment classification into MSL or NML, these NA where not taken into account. In this scenario,
if a locus for 10 samples, under the uninformative option, had 9 0/0 and 1 NAs then it was
missclassified as MSL beacuse it was calculating frequency of methylation states as 1-9/10=0.1
(>0.05 threshold, MSL), when it should be 1-9/9=0 (<0.05, NML).

This missclassifation was also causing that were, when transformed, MSL with all NAs, causing
that polymorphic() yielded NULL values, crashing the run (as reported first by Rik Lievers).

Changing the denominator in the calculation of methylation states frequency in methStatusEval.R
solved the problem.

Fix #10